requirements.txt
# Core
streamlit>=1.35.0
python-dotenv>=1.0.0

# LLM
openai>=1.30.0

# RAG
sentence-transformers>=3.0.0
pypdf>=4.0.0
numpy>=1.26.0

# Web Search
requests>=2.31.0
duckduckgo-search>=6.1.0
